The most significant changes in the code include updates to the AES, CRC, MultipleRetryFailureException, MD5, PluginHelper, ProcessUtils, PushQueue, RemainBytes, SignalToDigitalProcessor, StringByteUtils, Timer classes, and the Utils.csproj and UtilsUnitTest.csproj files. These changes primarily involve refactoring to improve readability, remove duplicate code, and update exception handling. Additionally, the version number in Utils.csproj has been increased and new test methods have been added in StringUtilsTest and TimerTest classes.
1. The AES class in AES.cs has been updated to include more specific ArgumentNullExceptions and to dispose of the _encryptor and _decryptor objects in the Dispose method.
2. The CRC class in CRC.cs has been refactored to remove duplicate code and improve readability. The Crc16, GBcrc16, HBcrc16, and CRC16_R methods now call a new private method, ComputeCrc16, to perform the common logic.
3. The MultipleRetryFailureException class in MultipleRetryFailureException.cs has been updated to allow for nullable inner exceptions.
4. The MD5 class in MD5.cs has been updated to use the using statement for the MD5 object, and to use the Equals method for string comparison in the VerifyMD5Hash method.
5. The PluginHelper class in PluginHelper.cs has been refactored to remove duplicate code and improve readability. The GetChildren methods now call a new private method, GetChildren, to perform the common logic.
6. The ProcessUtils class in ProcessUtils.cs has been updated to use the CancellationToken structure instead of the CancellationTokenSource class, and to throw an OperationCanceledException instead of a TaskCanceledException.
7. The PushQueue class in PushQueue.cs has been updated to use the CancellationToken structure instead of the CancellationTokenSource class, and to use the CancellationToken in the Task.Delay method.
8. The RemainBytes class in RemainBytes.cs has been updated to make the Bytes, StartIndex, and Count properties read-only, and to simplify the NextIndex and AvaliableSize properties.
9. The SignalToDigitalProcessor class in SignalToDigitalProcessor.cs has been updated to pre-calculate the scale factors in the constructor, and to use these pre-calculated values in the InputDataProcess and OutputDataProcess methods.
10. The StringByteUtils class in StringByteUtils.cs has been updated to simplify the StringToBytes and ComibeByteArray methods.
11. The Timer class in Timer.cs has been updated to use the CancellationToken structure instead of the CancellationTokenSource class, and to use the CancellationToken in the Task.Delay method.
12. The Utils.csproj file has been updated to increase the version number from 8.0.6 to 8.0.7.
13. The StringUtilsTest class in StringUtilsTest.cs has been updated to add new test methods for the MD5, Crc, and ComibeByteArray methods.
14. The TimerTest class in TimerTest.cs has been updated to add a new test method for the StopAsync method.
15. The changes in the UtilsUnitTest.csproj file are primarily related to formatting and the addition of a new property in the PropertyGroup section. The LangVersion line was added to set the version of the C# language to be used in the project to the latest available version. The rest of the changes are related to the formatting of the XML.

+++ Utils/StringByteUtils.cs
@@ -35,7 +35,7 @@ public static class StringByteUtils
         s = s.Replace(" ", "");
         byte[] buffer = new byte[s.Length / 2];
         for (int i = 0; i < s.Length; i += 2)
-            buffer[i / 2] = (byte)Convert.ToByte(s.Substring(i, 2), 16);
+            buffer[i / 2] = Convert.ToByte(s.Substring(i, 2), 16);
         return buffer;
     }
 
@@ -68,6 +68,7 @@ public static class StringByteUtils
         }
         return rs;
     }
+
     public static double ToDouble(byte[] value, int startIndex, bool isHighByteBefore = false)
     {
         var temp = value;
@@ -188,6 +189,7 @@ public static class StringByteUtils
         }
         return BitConverter.ToUInt64(temp, startIndex);
     }
+
     public static byte[] GetBytes(short value, bool isHighByteBefore = false)
     {
         if (isHighByteBefore)
